"10:05am. https://www.reddit.com/r/ProgrammingLanguages/comments/f29lxx/a_few_questions_for_compiler_engineers_out_there/
For some reason this thread got erased by the OP, but /u/jdh30 wrote some interesting things.

///

> Just out of curiosity, how did you manage to go from a dynamic language like Mathematica to a static one like F#? What would the reason be for doing something like this?

Their code base was extremely weird due to decades of organic evolution with no technical architecture or even direction. Firstly, their workflow was to have actuaries define calculations in Excel or Word and give them to one team who coded them up in Mathematica with initial tests. Then another team would translate the Mathematica to C++ to go into production and they would retest using a larger battery of tests.

In practice, translating Mathematica to C++ was extremely hard. They solved this problem by massively restricting the subset of Mathematica they used to the point that it was basically C++ code with a nicer syntax. A side effect of the consistent failures in manual translation was the evolution of an absolutely enormous test suite. By far the largest test suite I've ever seen. When I started, running the test suite in Mathematica took several days.

So their Mathematica code was already using an almost-entirely-sane type system, albeit a manual one, and I could easily and confidently test my output code by running the huge test suite through it.

> If I had to do something like that myself, the first thing that comes to mind is that I'd write a Mathematica interpreter in F# and point it at that code...but I am not sure that this would improve the situation any.

They hired me because I'm the only person in the world to have worked on both the Mathematica and F# language implementations before. And writing and using an interpreter proved to be extremely useful in this case because they had paid WRI a £2m consultancy contract to develop a bespoke Mathematica dialect for them that only ever worked with Mathematica 4 (1999) and could not be run in any more modern version of Mathematica. Rather than paying WRI £2m every time they released a new version of Mathematica, this client just stuck with Mathematica 4.

So the interpreter was actually very useful for several reasons:

Nothing else could even parse their code because it was a custom dialect of a prioprietary language.

My interpreter could run their huge test suite as a proof of concept, showing them the feasibility of the much larger transpiler project.

The execution of the test suite verified my inferred types across their Mathematica code.

So it was actually surprisingly valuable. My interpreter also ran their test suite a lot faster than Mathematica did, bringing the time from days to hours. The final F# code ran the entire test suite in seconds.

The overall benefits for the client were:

1. Much easier to hire/train F# developers.

2. Much cheaper to use F#.

3. 5x higher developer productivity.

4. 10x less code.

5. 100,000x faster code.

6. Higher morale.

7. Easier to audit the algorithms.

///

I'll save this reply in my journal for posterity since I asked him those questions.

///

Forgive my rant but I think compiler engineering is a massively under-valued skill. In many companies over the decades I have noticed that management evolve an inefficient system where the shipping of a business idea from a business person to the customer has to go through IT. A vastly better solution is to pull IT out of that (where they are generally drowning in BAU projects anyway) and get IT to design and build bespoke tools that allow the business people to ship their own ideas. IT are generally happier doing this more interesting work. The business people are happier not having such frustrating interactions with IT. The company enjoys faster time to market and lower overheads. In my experience, the tool that IT builds for the business person is often a language implementation of some form.

The problem I see with compiler engineering is excessive focus on what, for most projects, is massive overengineering. I think the minimum viable product gets neglected.

///

Let me also save this rant from a previous post. I'll keep this bit of advice in mind.

10:35am. I've fought off my grogginess a little and got some morning reading done. Let me go through the TS documentation for a while. There is also the Playground with code examples.

https://www.typescriptlang.org/docs/home.html

11:10am. https://distill.pub/2020/growing-ca/
Growing Neural Cellular Automata

This is interesting. I'll read it later.

11:30am. https://www.typescriptlang.org/docs/handbook/interfaces.html

The handbook is better than that course as a source of info.

I'll go through this today and then finally start coding. I really want to make some progress on this, but there is no helping it. Starting new things always requires a ramp up period.

12pm. https://www.typescriptlang.org/docs/handbook/classes.html

I like the documentation much more than the material I went through yesterday. I am starting to acclimate to this.

12:05pm. I am actually starting to get interested in TS a little. It has pattern matching and all that good stuff. For a static language, it has features I've only ever used in Spiral in various guises.

...Enough. Let me stop here for a while so I can eat.

12:10pm. Without a doubt, I will finish the documentation today and get some real programming done after that."