"5:05pm. Solo Leveling is out. Let me read that and then I will resume. I am starting to find my center again.
5:10pm. That was fun. Let me put in the last bit of work for the day.

I definitely do not feel like programming, but what I do feel is getting that example to run. Right here is where I should be making the first step. Here I need to do icebreaking more than anything else.

https://www.youtube.com/watch?v=4tk0Ak-dEjs

Let me go through that tutorial again.

5:30pm. Ok, I installed those generators and yarn. ESLint as well. I remember hearing about that tool in one of the talks years ago.

5:40pm. Let me stop here for the day. I just do not feel like going at it for any longer.

I have to follow a lot of instructions in that video, and I do not have the focus for this anymore.

5:45pm. I'll do it properly starting tomorrow. What I did now with installing those things is good enough as icebreaking.

In fact, from installing TS, to this now, everything is going super smoothly. Last time...remember when I tried the F# plugin how broken everything was? I was really afraid how things would go given how wild west the JS ecosystem is, but I should not have been so scared.

The way things are going now is not so bad. I spent 3 days on getting familiar with TS, and now I am going to start getting things off the ground.

5:55pm. I can do this.

Give me a week and I will figure out all the prerequisites to editor support. After that is done, I will be able to leave this aside to work on the typechecker. That is where the real fun will start."