"12:35pm. Done with breakfast, done with chores. Let me start.
As expected it is one thing to be motivated to do great things, and it is quite another to do the grind.

Seriously, Spiral v0.2 is the last time I will subject myself to this. I am really unfortunate that nobody else made a language like Spiral, and that my talents are well suited for this kind of design work.

But being in this room, trying to painstakingly put the pieces together is...inhuman.

It was one thing to do this in 2017, but 2020 is really my limit. I want to be done with the most of it here, so that I can do more interesting programming after that.

In the future I'll either want to offload this work to AI agents, or find a way to literally mind control myself into enjoying this properly.

There is only so many times I can play the willpower card.

1:05pm. https://code.visualstudio.com/api/extension-guides/virtual-documents

I can't really say I am studying this. At best, I am dwelling in my thoughts while sampling the documentation on the side.

Let me do that for the whole API docs and then I will go directly for the samples.

1:30pm. This kind of study is hugely wasteful of time and it reminds me a bit of my time in school when the teacher will drone on and I would be lost in my dreamworld, not at all paying attention to the lecture.

Nonetheless, it is the best I can do. I guess it is too much to expect that I'll be able to finish this in a week. VS Code plugins are APIs piled on APIs and I can't really grok that just like that. I am simply going to have to do it like this, in multiple passes, each time internalizing something new.

If there is one thing I am grateful, it is that I am not checking social media and Mangdadex every 15m like yesterday.

1:35pm. https://code.visualstudio.com/api/language-extensions/syntax-highlight-guide

Well, let me just continue at my current pace. Since I am already more than halfway into it, I'll give this documentation another pass. Then I'll just focus on studying the samples.

1:40pm. Focus me, focus. Do not get lost in thought.

Actually, this journal is a problem on its own. It drew me out of the zone. I'll try to stop writing it until I am at least done with the docs.

https://www.apeth.com/nonblog/stories/textmatebundle.html

Let me read this. I have no idea what Textmate grammars are. The VS Code documentation is the first time I've encountered them.

1:55pm. Ok, I skimmed that. Do I really have to use Textmate? Can't I use my own parser?

2:05pm. https://code.visualstudio.com/api/language-extensions/syntax-highlight-guide#contributing-a-basic-grammar

Ok, I have to focus on this just a little. I can do keyword coloring for Spiral with this.

In fact, that should be the very first thing I do, as a part of the project.

2:30pm. Done with that section. I will keep it in mind. Obviously I do not get it all just from reading this, but through experimentation I should be able to figure it out.

2:50pm. https://code.visualstudio.com/api/language-extensions/language-server-extension-guide

Nearly done with docs. I need a short break here. After that I will finish this up and move to the samples.

3:10pm. Let me resume.

Let me finish this last bit of the docs and then I will move to the samples. Reading the docs is like reading a math book. It is not real skill - real skill is only acquired once you start playing around with the stuff yourself and put some of that into procedural memory.

3:15pm. Oh, Kengan Omega is out. Let me read this first then. I've already been spoilered that Ohma comes comes back.

3:35pm. I had no idea I would relax for so long. The chapter was hype. But seriously, let me resume.

3:55pm. Now I started reading threads on the ML sub. Focus me, focus!

4pm. https://code.visualstudio.com/api/language-extensions/language-server-extension-guide

I have to admit, there are a shitload of details in here.

I said I wasted the morning, but it feels like I wasted everything after that. In the morning I at least did that tutorial. Let me just finish this and then I will finally move to the samples.

All this reminds me when in 2015 I spend like a month wastefully studying GUIs. I autistically went through all the examples in Petzold's book like it was a sacred tome.

4:05pm.

```
{
  "name": "Language Server E2E Test",
  "type": "extensionHost",
  "request": "launch",
  "runtimeExecutable": "${execPath}",
  "args": [
    "--extensionDevelopmentPath=${workspaceRoot}",
    "--extensionTestsPath=${workspaceRoot}/client/out/test",
    "${workspaceRoot}/client/testFixture"
  ],
  "stopOnEntry": false,
  "sourceMaps": true,
  "outFiles": ["${workspaceRoot}/client/out/test/**/*.js"]
}
```

Yesterday, I said I am starting to warm up to Typescript, but so far literally everything is like this. This is not like learning algorithms. Everything is one huge DSL, built out of smaller ones. All the functions have optional parameters, I literally cannot even begin guessing how complex all this is out of the hood or how it fits together.

4:15pm. Done with the documentation. I've skimmed it once more. I really did not get anything towards the end, and now the tedium is really hitting me. I am not sure whether I am up for playing with samples like I said I would, but let me give it a try.

https://github.com/microsoft/vscode-extension-samples

I think I'll start things off by cloning this.

4:25pm. I'll give the status bar example a try.

...Why is it building so long? This is ridiculous. Maybe I should have ran it in debug mode.

4:35pm. No, it seems like I am doing wrong. Running F5 is not starting another instance of code with the plugin active. I have no idea what is going on.

Let me do one thing. There are two other plugin tutorials on Youtube so let me watch those.

4:50pm. https://youtu.be/srwsnNhiqv8

I am trying to watch this, but my focus is really dead at this point.

I should be figuring out how to get that status bar extension to run, but I do not feel like it. Seriously, fuck documentation. I am such a masochist.

Tomorrow, I am going to go into this more creatively if possible.

The way I've been approaching it so far is the worst.

All this really makes me appreciate programming though. If this was math, I'd have to basically throw in the towel here, but I know that with programming if I trace it, I can figure out everything.

That reminds me.

Remember my Pharo experience and how I managed to go by simply tracing through everything? I should be doing that here. I mean I have the debugger. I should use it.

I actually completely forgot about this before I mentioned tracing just now.

I thought Pharo was a waste of time, but I did get something from it in the end.

I should be tracing, tracing, tracing the shit out of everything, especially in the beginning.

Things like tracing and randomized testing is something math lost when it detached itself from a computational basis.

5pm. What I did in the last week is something I should not make a habit of.

This spending of time is a great show of will, but not so much of intelligence. Pharo really taught me that you should just take courage and dive into the system if you want to understand it.

If I am going to master these examples, I should channel the experience of last year.

Let me stop here for the day. Rather than trying to push myself, I should just take a bath instead and do this trace when I am fresh and ready to go."