"1:10pm. https://www.quantamagazine.org/the-map-of-mathematics-20200213/
I'll keep this article for later. I am done with breakfast here.

Let me just slack a little, do the chores and then I will resume.

1:55pm. It seems I will be doing some housework soon, so let me hold off resuming for a bit.

2pm. https://github.com/Microsoft/vscode-extension-samples/tree/master/snippet-sample

Ah, what the hell, let me just do it while I wait. Let me take a look at the snippet sample.

https://code.visualstudio.com/docs/editor/userdefinedsnippets#_creating-your-own-snippets

Ah, I hadn't known about this. I was surprised to find only a snippets.json file in that project.

2:15pm. https://github.com/Microsoft/vscode-extension-samples/tree/master/language-configuration-sample

Forget snippets. It might be interesting to keep in mind, but I am not into copy paste.

Let me check out this next.

2:20pm. https://code.visualstudio.com/api/language-extensions/language-configuration-guide

Let me save here. I need to check out that second round of chores.

2:35pm. The schedule of homework seems to be intermittent so let me put some study in while I wait.

```
{
 "comments": {
  "lineComment": "//",
  "blockComment": [ "/*", "*/" ]
 },
 "brackets": [
  ["{", "}"],
  ["[", "]"],
  ["(", ")"]
 ],
 "autoClosingPairs": [
  { "open": "{", "close": "}" },
  { "open": "[", "close": "]" },
  { "open": "(", "close": ")" },
  { "open": "'", "close": "'", "notIn": ["string", "comment"] },
  { "open": "\"", "close": "\"", "notIn": ["string"] },
  { "open": "`", "close": "`", "notIn": ["string", "comment"] },
  { "open": "/**", "close": " */", "notIn": ["string"] }
 ],
 "autoCloseBefore": ";:.,=}])>` \n\t",
 "surroundingPairs": [
  ["{", "}"],
  ["[", "]"],
  ["(", ")"],
  ["'", "'"],
  ["\"", "\""],
  ["`", "`"]
 ],
 "folding": {
  "markers": {
   "start": "^\\s*//\\s*#?region\\b",
   "end": "^\\s*//\\s*#?endregion\\b"
  }
 },
 "wordPattern": "(-?\\d*\\.\\d\\w*)|([^\\`\\~\\!\\@\\#\\%\\^\\&\\*\\(\\)\\-\\=\\+\\[\\{\\]\\}\\\\\\|\\;\\:\\'\\\"\\,\\.\\<\\>\\/\\?\\s]+)",
 "indentationRules": {
  "increaseIndentPattern": "^((?!.*?\\/\\*).*\\*\/)?\\s*[\\}\\]].*$",
  "decreaseIndentPattern": "^((?!\\/\\/).)*(\\{[^}\"'`]*|\\([^)\"'`]*|\\[[^\\]\"'`]*)$"
 }
}
```

This is interesting. It seems like it could be useful, but hell if I like regexes. I went through that textmate thing two days ago, but I really like parser combinators the most.

5:15pm. Ok, I am finally done with chores just now. My brothers 1h can easily turn into 3.5.

I want to study the LSP example just a little before I call it a day.

5:20pm. https://github.com/Microsoft/vscode-extension-samples/tree/master/lsp-sample
> Type `j` or `t` to see `Javascript` and `TypeScript` completion.

I have no idea what this is supposed to be doing.

5:25pm. Nope. The VS Code instance is running fine, but it is not doing anything.

5:30pm. And my reaction to this is to start surfing 4chan. Damn it.

```
t
aaa bb cc AAA
```

Ahhh, I see. It gives me a warning on `AAA` that it is all uppercase.

5:40pm.

```
async function validateTextDocument(textDocument: TextDocument): Promise<void> {
 // In this simple example we get the settings for every validate run.
 let settings = await getDocumentSettings(textDocument.uri);

 // The validator creates diagnostics for all uppercase words length 2 and more
 let text = textDocument.getText();
 let pattern = /\b[A-Z]{2,}\b/g;
 let m: RegExpExecArray | null;

 let problems = 0;
 let diagnostics: Diagnostic[] = [];
 while ((m = pattern.exec(text)) && problems < settings.maxNumberOfProblems) {
  problems++;
  let diagnostic: Diagnostic = {
   severity: DiagnosticSeverity.Warning,
   range: {
    start: textDocument.positionAt(m.index),
    end: textDocument.positionAt(m.index + m[0].length)
   },
   message: `${m[0]} is all uppercase.`,
   source: 'ex'
  };
  if (hasDiagnosticRelatedInformationCapability) {
   diagnostic.relatedInformation = [
    {
     location: {
      uri: textDocument.uri,
      range: Object.assign({}, diagnostic.range)
     },
     message: 'Spelling matters'
    },
    {
     location: {
      uri: textDocument.uri,
      range: Object.assign({}, diagnostic.range)
     },
     message: 'Particularly for names'
    }
   ];
  }
  diagnostics.push(diagnostic);
 }

 // Send the computed diagnostics to VSCode.
 connection.sendDiagnostics({ uri: textDocument.uri, diagnostics });
}
```

This is where the actual errors are calculated.

```
range: Object.assign({}, diagnostic.range)
```

Is it copying the range here for some reason.

5:45pm. Ok...to me personally that particular function is the most important. It is the pivotal point in the whole example.

6:10pm. I've gone top to bottom over the server.ts file a few times now. I am definitely going to be going through this in depth in the future.

Let me take a look at the client file.

6:20pm.

```
 // The server is implemented in node
 let serverModule = context.asAbsolutePath(
  path.join('server', 'out', 'server.js')
 );
```

So it takes the path of the server directly.

```
 let serverOptions: ServerOptions = {
  run: { module: serverModule, transport: TransportKind.ipc },
  debug: {
   module: serverModule,
   transport: TransportKind.ipc,
   options: debugOptions
  }
 };
```

This is passed into here.

```
 client = new LanguageClient(
  'languageServerExample',
  'Language Server Example',
  serverOptions,
  clientOptions
 );
```

This is how it gets started.

This is about most of the client. Nice. Quite nice.

I can work with this. Obviously, I do not understand the specifics of this, but I am definitely going learn all of this by heart in the next month or so.

Let me stop here.

https://code.visualstudio.com/api/extension-guides/overview

There are still two more samples to study, and after that I'll fire up the tracer and go through this more in depth.

6:25pm. There is no mistake, mastering editor support will be just like learning GUIs all over again. It is going to take a while. It is not a trivial thing by any means. But it is definitely doable with effort. And it is definitely worth the effort.

GUIs are ultimately what separate serious programs from toys, and editor support is what separates serious languages from throwaway academic experiments.

Providing editor support for Spiral will be my way of saying to the world that it is ready.

I definitely admire F#'s great VS integration. It just makes everything so much better. Imagine if I programmed in it the same way as Spiral in 2018 - it would be horrible. I am not going to let it end on that note.

I am going to do much better this time."