"8:50am. I am up. I think I have a cold, but it is nothing serious.
Let me do some morning chilling and then I will go through the other two LSP examples.

https://boards.4channel.org/sci/thread/11387295

I am amazed that there is an optimization thread on /sci/.

9:25am. Ok, enough slacking. I want to take a look at the last two examples. I'll think of what comes next after that.

9:40am. Focus me, focus.

https://github.com/Microsoft/vscode-extension-samples/tree/master/lsp-log-streaming-sample

9:45am. Ok, now I have a problem that the example does not run.

```
"languageServerExample.trace.server": {
  "format": "json",
  "verbosity": "verbose"
},
```

Also where am I supposed to add this setting? Let me check out the JSON files.

```
    "languageServerExample.trace.server": {
     "scope": "window",
     "type": "object",
```

There is this is `package.json`. So maybe it wants me to overwrite it?

9:50am. Well, forget his thing for now. Let me report it in the issue. After that I'll take a look at the code itself.

https://www.w3schools.com/js/js_operators.asp

```
 hasConfigurationCapability =
  !!capabilities.workspace && !!capabilities.workspace.configuration;
```

Hmmm...I thought that !! would be the not operator or something like that, but...

Wait, `&&` is definitely `and` based on the table. Does that mean that `!!` is `not not`? What the hell?

https://stackoverflow.com/questions/784929/what-is-the-not-not-operator-in-javascript

10:25am.

```
while ((m = pattern.exec(text)) && problems < settings.maxNumberOfProblems) {
```

How does `m = pattern.exec(text)` work. I am going to assume that it returns `m` after doing assignment which is evaled as `true` if it is not `null`.

10:35am. Done with the example. Let me move to the last one. Then I will start thinking about the next phase.

10:40am. Focus me, focus.

10:55am. The last example is particularly simple as it does nothing in particular.

With that let me stop here for a bit.

In summary, I am really disgusted what a waste of time the past week was and how inefficiently I am going through this. My Pharo experience was much better than this. It is making me realize how helpful being able to trace through the system really is. This is garbage.

At any rate, that is what I am going to do next.

All this effort of the past week was just there so I can get some basic familiarity with the system, so that I do not freeze like a deer in headlights when the fight starts.

11:05am. Ok...

From here on out the plan is simple...go through that LSP example and trace through it. Then do that exercise for some real plugin.

Then start work on my own.

Once I have the scaffolding down, I can resume work on Spiral. I really am grateful that I am going to have to master only LSP, and not the whole system. That will make this whole affair much eaiser."