"12:30pm. A bit more resting, the chores and then I will resume.
1:50pm. Done with chores.

Since Reddit since to be down, I can't see whether I got any answer to my PL sub question. That means I should go with my plan to try other plugins. Let me check out that page with the list of them.

2pm. Damn it, where are the existing plugins? Where was the page for them?

2:05pm. https://microsoft.github.io/language-server-protocol/implementors/servers/

I have no idea where I found the link to this, but it is good.

2:15pm. It does not seem like there are any C specific plugins. It did not occur to me earlier, but maybe the language that most closely meets my simplicity requirements would be Go.

https://github.com/fsharp/FsAutoComplete

Still, another F# plugin exists. I thought that the other would be Ionide, but that was not it.

Ok...here is what I am going to do. Let me give this plugin a try. Then I will try the C# one.

This time I won't be ripping myself trying to study this and to get it to work.

Had I known things would have been like this, I'd have changed my focus earlier.

2:20pm. Obviously the primary thing I need to do is to just get things to work in the first place. Then I can pick a study target.

Ok, focus me.

Let me clone that thing and I will give it a try. If it starts giving me bullshit like the rest, I will not waste too much time trying to get it to work.

2:30pm. https://github.com/fsharp/FsAutoComplete

The instructions here seem to be working. I can't do it directly from the solution though. Nonetheless, if I am lucky I might be able to run it from the IDE. Right now the packages are being downloaded.

2:35pm. Ok, the thing finally built. Just how many dependencies does the damn thing need?

At any rate, let me give it a try.

Ok, the thing compiled, but how do I run it?

...Well, the exe is there. I can't build it from the solution now though. Nor run it. What are the devs of this using?

I have no idea.

Also FsAutoComplete is enormous. I can't deal with this? Forget it.

What is the next step?

2:50pm. https://github.com/OmniSharp/csharp-language-server-protocol

10 days ago I did some Omnisharp stuff. I completely forgot about that.

https://github.com/joelday/papyrus-lang

Something like this might be up my alley.

...Hmmm. I'll go with the V# LSP first. Maybe I will have more like with that than the F# stuff.

2:55pm. This thing is failing to download packages from Nuget as well. I don't get why this is happening.

Should I not be building things directly from the IDE?

The build is taking a while too. Let me give it another minute and I will cancel it if need be. I wish the IDE gave me some sort of status indication what it is doing.

3pm. Oh, it succeeded. Now what is something nice for once.

3:05pm. Ok, the thing runs. So this was more successful and painless than my past attempts combined. The C# plugin deserves a lot of praise for actually being easy to build and to run.

But as expected the complexity of this is off the scale.

3:15pm. It is too big. It is just too big. Projects at this scale are the programming equivalent of math straddling the infinity and proofs that pull things out of thing air.

3:25pm. I need a break as I am pretty discouraged right now. Nothing is really going right here.

I just do not know why. I get the feeling that going in this direction should work, but the impetus is just not there. Programming has an evolutionary aspect to it, and it is really hard to start from the middle. The same applies to math and all sciences in fact.

This is why small examples, principles and the ability to test one's ideas are so important.

3:30pm. There is no way around this. I am going to give that Papyrus plugin a try. If that turns out to be too complex as well, I am in for a rough ride. I suppose checking out Go after that is an option, but the right choice in that case would be to learn some webdev for real.

At least enough to set up my own servers and such.

3:35pm. It can't be helped, the world is really harsh.

One more time. Let me roll up my sleeves and see where that gets me.

3:40pm. Nope, the build fails. This makes things easy for me to drop it.

I am not going to bother to look for reasons why this is happening.

http://www.omnisharp.net/

Once more, since just like the C# LSP the Papyrus uses Omnisharp let me take at the docs for this.

3:45pm. ...It does not have any docs.

3:50pm. I am really disappointed right now. I am disappointed to the limit.

I do not really feel like doing anything else right now.

I think it fair to say that plan A is a bust.

3:55pm. The main plan was to find something good and trace through it in order to attain understanding similar to Pharo, but what is going on here is super underwhelming. I am extremely uninspired by the code that I've seen.

4pm. Plan B is to learn webdev from the ground up.

Once I level those skills up enough, I should have enough knowledge to understand whatever is it that the VS samples are trying to tell me.

If it is merely understanding how to set up a client, I think I have enough knowledge the in the first F# plugin that I've studied.

But setting up an actual server is something else. Even though the idea how to do it is dead simple.

4:05pm. The main thing I am missing is the minimal example.

It is one thing to be given a mountain of a code. And it is one thing to have a completed editor.

But what is the ideal server? I can't figure out where F# and C# went wrong here.

4:10pm. Without a base, it is possible to come into a big project and write something up based on the requirements given to you, but it is not possible to improve the whole architecture.

The base is the answer to the question of not just what they are, but what the assumptions should be.

4:15pm. Creating web servers has an immense amount of overlap with language servers.

I can't avoid it anymore. Life it forcing me to take responsibility. I have to learn webdev to some degree here. I have to learn the basis for it, the backend stuff. Typescript should serve fine for this.

After that is done, I should not have much trouble figuring out how to create a server in F#.

4:20pm. Really, the way the language serve will work is super simple. It will just redo the whole module (and the ones after it on focus) every time the text changes. As long as typechecking is fast, this will not be an issue.

Let me take a little break here.

Remember those 3 Typescript books I got a while ago from Genesis? I got them exactly for this occasion. It is time to start seriously gathering motivation for the part that will come."