"9:40am. Got up 10m ago and am just chilling.
Let me do my reading and then I will start. I haven't gotten much in terms of replies on the PL sub, so from that I am sure that I am not well grounded at the moment. It is not my imagination that this is overly difficult. And the reason for that is my lack of webdev knowledge.

I've definitely looking down too much on webdev. Sure, JS is shit, but ultimately webdev is about communication. Communicating processes, communicating nodes, communicating servers...web shit is what floats at the top. At the bottom is what I actually need to understand which is asynchronous communication.

10:20am. Ok, enough slacking. Though I feel like just gathering my will rather than diving into the book right away.

Let me just say, I really botched the last ten days. I've been really stubborn on finding a good template. I've been resolved to make the plugin, but I had not been resolved to necessarily dive into the far end of webdev.

But I should have known, though servers are used for web, they do not have to be in any language specifically. It is just that the context they are in clearly highlights their need. So that is where they came into being.

10:30am. I've been resolved to end it in a week or two, but now that I've failed I am resolved to have this take months. That is what failure usually looks like for me.

10:35am. Ahhh...I do not feel like going through this quickly. I guess Spiral will get done when it wants to be done. What I will do now is go back to 2015, and turn on the study mode again.

10:40am. Enough dwelling on this. Let me give chapter one a try again.

Oh yes. It occurred to me during the night, but I forgot about it till now. What I will be studying here will be related to the interface creation mission once I have the agents ready. I find it ironic that I am doing this out of order, but nevertheless it is good that I am claiming these hidden benefits.

10:45am. Hmmm, VS Code is nearly 3Gb when fully unpacked on disk.

That is a lot of dependencies.

10:50am. Hmmm...I think I did a minimalist install of VS with just F# support. I think I'll have to get Typescript separately. Let me do so now.

11am. Ok, I've installed the typescript support for VS 2019, but where are the templates?

11:05am. Ah screw it. Though it is 3Gb let me get full webdev support.

I might as well take that break here.

11:50am. Let me put in a few words even though it is my break since I forgot to commit earlier. The thing installed and I see the ASP.NET template. There is such a template for F# as well, so I'll rather go with that than C#. Great. It is a nice touch from the devs to put the `new` icon to indicate the new additions to the VS instance.

There is nothing under Typescript itself for some reason. Well, the devs probably put it there as on option."