"8:35am. Got up 20m ago. Yesterday fatigue really hit me like a truck, so I did not even manage to read more than 2 chapters of the Toji novel. I'll take the opportunity provided by this early awakening to indulge myself. I'll study the ASP.NET book after that.
Also, I did some planning. I've decided. I think it really is unrealistic to aim for low 6-figure remote job at Intel, but I should be able to use that in order to pivot and get the guy to send me a sample. Having that neurochip would be the catalyst to actually doing the RL agent properly. If I could have it, I am not sure how things would go from there, but I am certain that I could at least create the poker agent. That would matter a lot.

If I get an offer for a 5 figure remote job, it could also serve as a backup if any disasters strike even if I have no intention of taking it. I'll be killing many birds in with one stone this way. The salary money does not really matter, all the big stuff in the lead up to the Singularity will be through RL anyway.

9:25am. https://news.ycombinator.com/item?id=22381129
"Location: Remote
Vathys is hiring a machine learning compiler engineer to build the core compiler to a system that will be a step function improvement in ML capability.

The ideal candidate:

1. Has a good understanding of the fundamentals of deep learning algorithms. Less "I can use Keras" and more "I can write a CNN from scratch in numpy".

2. Has used a NN compiler framework or IR framework in the past such as TVM, MMdnn, MLIR, Glow, Tensor Comprehensions, etc.

3. Has experience with compilers, especially optimizing compilers and compilers for high performance code.

4. Has experience writing kernels for deep learning applications.

5. Has at least passing familiarity with the concept of polyhedral optimizations

Generous compensation will be provided to the right candidates.

Please contact tapabrata_ghosh[at]vathys[dot]ai if you're interested"

They forgot to put a link to their site, http://vathys.ai/.

It seems this is one of the startups, but damn it, it is just too early for me. I really do need to finish Spiral before I apply to anywhere, but on the other hand if I wait too long the opportunities will go away.

Nonetheless, times like these are a test of character. I always need to be in the frame of mind of abundance.

There will be more of these in the future. Plus, I am not making Spiral for any specific piece of hardware. There is no point in committing myself anything until I know that I can use it for my own purposes. I need to keep in mind that most of these startups are here to drain investor money and capitalize on the hype. There isn't room for 100 different neurochips on the market.

10:20am. Done with chapter 4 of the Toji novel. That was really good. I enjoyed this morning.

Now let me study the ASP.NET book."