"3:40pm. It was super easy to open that Azure DevOps account.
Oh, I see that the poker thing I did back in 2016 is still there in the private repo. I actually forgot that the F# ML library that I did was called Spiral.

I've been thinking during the off-time whether I want to continue with my plan. Since I am firmly on the path of learning webdev, rather than going with the banditry plan, what I should do instead after training up that poker agent is distill it and turn the whole thing into a game.

It is just the case, that making the language, training the agent and everything else is difficult enough on its own. If I have to fight the online poker sites in addition to all of that, I will have a hard time. Even if I have the agent, I do not have a good way of automating the stealth Turing tests the sites would set out for me.

They can be beaten with hand engineering, but that would be a useless waste of time I could use making longer term gains and learning new skills.

I'll crush them when I am good and ready, but for now why don't I go this way and aim to diversify my portfolio. I might as well try capitalism for real at least once before humanity gets driven to extinction by the AI whose creation I am contributing to and whose power I am trying to internalize.

4:45pm. 93/558. It is teaching me how to use Github here. Some of this is new to me, so I've been going through this slowly.

I admit, this book is more thorough in starting me out than I thought it would be.

5pm. 97/558. I hadn't realized that it is possible to connect to a project.

5:10pm. I am looking for the build option, and it seems it is under repos rather than under pipelines unlike in the book.

5:15pm. Er, no, I actually just needed the classic editor. I see it now.

5:30pm. "In the next two chapters, we will explain the basic concepts of ASP.NET Core 3, including the start up class, using middleware, routing, error handling, and much more."

I saw a lot of this in the plugins so I am definitely interested in the meat of the book.

At any rate, I am finally done with chapter two. A lot of the stuff here is new to me, but it was not very interesting. As the project is empty, the build obviously failed.

5:35pm. 103/558. Do I really feel like continuing here?

5:40pm. Though my focus was low, I did go through 70 pages today. It was tedious tutorial stuff, but effort is still effort. I am feeling the fatigue.

Let me just start the chapter and I'll leave it for tomorrow. Hopefully the book should start getting interesting here. I want to save the good for when I am fresh.

6pm. 106/558. I'll leave this for tomorrow.

All in all, it does not feel like I did too badly today. Yes, I wasted a lot of time and my focus was low, but the material itself was tedious. Still I did learn a bit about modern programming practices. Tomorrow will be more hands on, much like yesterday.

6:10pm. Now come leisure. A best practice is to have learning and rest in balance. Stress maintenance is a core skill unless one wants frequent trips to the doctor."